pyproject.toml
[tool.poetry]
name = "rest-api"
version = "0.1.0"
description = ""
authors = ["sajad <fallahdoostsajad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
